Add test for createTempFile
https://jira.i-novus.ru/browse/RDM-1175

+++ rdm-impl/src/test/java/ru/i_novus/ms/rdm/impl/file/TempFileTest.java
@@ -45,22 +45,39 @@ public class TempFileTest {
     private void testDirectory(String name, File directory) {
 
         log.info("Directory: {}={}", name, directory);
+        assertDirectory(directory);
 
-        assertNotNull(directory);
+        testCreateFile(directory);
+        testCreateTempFile(directory);
+    }
 
-        assertTrue(directory.exists());
-        assertTrue(directory.isDirectory());
-        assertTrue(directory.canRead());
-        assertTrue(directory.canWrite());
+    private void testCreateFile(File directory) {
 
         final File file = new File(directory, "testfile.ext");
         try {
             createFile(file);
+            assertFile(file);
 
         } finally {
-            if (file.exists()) {
-                deleteFile(file);
-            }
+            deleteFile(file);
+        }
+    }
+
+    private void testCreateTempFile(File directory) {
+
+        File file = null;
+        try {
+            file = File.createTempFile("tempfile", ".ext", directory);
+            log.info("Temp file '{}' created successfully", file);
+
+            assertFile(file);
+
+        } catch (IOException e) {
+            log.error("Error creating temp file '{}':\n{}", file, e.getMessage());
+            throw new RuntimeException(e);
+
+        } finally {
+            deleteFile(file);
         }
     }
 
@@ -81,10 +98,33 @@ public class TempFileTest {
 
     private void deleteFile(File file) {
 
+        if (file == null || !file.exists())
+            return;
+
         final boolean deleted = file.delete();
         if (deleted)
             log.info("File '{}' deleted successfully", file);
         else
             log.info("File '{}' cannot be deleted", file);
+    }
+
+    private static void assertFile(File file) {
+
+        assertNotNull(file);
+
+        assertTrue(file.exists());
+        assertTrue(file.isFile());
+        assertTrue(file.canRead());
+        assertTrue(file.canWrite());
+    }
+
+    private static void assertDirectory(File directory) {
+
+        assertNotNull(directory);
+
+        assertTrue(directory.exists());
+        assertTrue(directory.isDirectory());
+        assertTrue(directory.canRead());
+        assertTrue(directory.canWrite());
     }
 }
